+++ src/textual/css/stylesheet.py
@@ -12,7 +12,6 @@ from rich.console import Console, ConsoleOptions, RenderableType, RenderResult
 from rich.markup import render
 from rich.padding import Padding
 from rich.panel import Panel
-from rich.style import Style
 from rich.syntax import Syntax
 from rich.text import Text
 
@@ -26,7 +25,7 @@ from .parse import parse
 from .styles import RulesMap, Styles
 from .tokenize import Token, tokenize_values
 from .tokenizer import TokenError
-from .types import Specificity3, Specificity6
+from .types import CSSLocation, Specificity3, Specificity6
 
 _DEFAULT_STYLES = Styles()
 
@@ -68,27 +67,34 @@ class StylesheetErrors:
         for token, message in errors:
             error_count += 1
 
-            if token.path:
-                path = Path(token.path)
-                filename = path.name
-            else:
-                path = None
-                filename = "<unknown>"
-
             if token.referenced_by:
                 line_idx, col_idx = token.referenced_by.location
             else:
                 line_idx, col_idx = token.location
             line_no, col_no = line_idx + 1, col_idx + 1
-            path_string = f"{path.absolute() if path else filename}:{line_no}:{col_no}"
-            link_style = Style(
-                link=f"file://{path.absolute()}" if path else None,
-                color="red",
-                bold=True,
-                italic=True,
+
+            display_path, widget_var = token.read_from
+            if display_path:
+                link_path = str(Path(display_path).absolute())
+                filename = Path(link_path).name
+            else:
+                link_path = ""
+                filename = "<unknown>"
+            # If we have a widget/variable from where the CSS was read, then line/column
+            # numbers are relative to the inline CSS and we'll display them next to the
+            # widget/variable.
+            # Otherwise, they're absolute positions in a TCSS file and we can show them
+            # next to the file path.
+            if widget_var:
+                path_string = link_path or filename
+                widget_string = f" in {widget_var}:{line_no}:{col_no}"
+            else:
+                path_string = f"{link_path or filename}:{line_no}:{col_no}"
+                widget_string = ""
+
+            title = Text.assemble(
+                "Error at ", path_string, widget_string, style="bold red"
             )
-            path_text = Text(path_string, style=link_style)
-            title = Text.assemble(Text("Error at ", style="bold red"), path_text)
             yield ""
             yield Panel(
                 self._get_snippet(
@@ -132,7 +138,7 @@ class Stylesheet:
         self._rules_map: dict[str, list[RuleSet]] | None = None
         self._variables = variables or {}
         self.__variable_tokens: dict[str, list[Token]] | None = None
-        self.source: dict[str, CssSource] = {}
+        self.source: dict[CSSLocation, CssSource] = {}
         self._require_parse = False
         self._invalid_css: set[str] = set()
         self._parse_cache: LRUCache[tuple, list[RuleSet]] = LRUCache(64)
@@ -202,7 +208,7 @@ class Stylesheet:
     def _parse_rules(
         self,
         css: str,
-        path: str | PurePath,
+        read_from: CSSLocation,
         is_default_rules: bool = False,
         tie_breaker: int = 0,
         scope: str = "",
@@ -211,7 +217,7 @@ class Stylesheet:
 
         Args:
             css: String containing Textual CSS.
-            path: Path to CSS or unique identifier
+            read_from: Original CSS location.
             is_default_rules: True if the rules we're extracting are
                 default (i.e. in Widget.DEFAULT_CSS) rules. False if they're from user defined CSS.
             scope: Scope of rules, or empty string for global scope.
@@ -222,7 +228,7 @@ class Stylesheet:
         Returns:
             List of RuleSets.
         """
-        cache_key = (css, path, is_default_rules, tie_breaker, scope)
+        cache_key = (css, read_from, is_default_rules, tie_breaker, scope)
         try:
             return self._parse_cache[cache_key]
         except KeyError:
@@ -232,7 +238,7 @@ class Stylesheet:
                 parse(
                     scope,
                     css,
-                    path,
+                    read_from,
                     variable_tokens=self._variable_tokens,
                     is_default_rules=is_default_rules,
                     tie_breaker=tie_breaker,
@@ -263,7 +269,7 @@ class Stylesheet:
             path = os.path.abspath(filename)
         except Exception:
             raise StylesheetError(f"unable to read CSS file {filename!r}") from None
-        self.source[str(path)] = CssSource(css, False, 0)
+        self.source[(str(path), "")] = CssSource(css, False, 0)
         self._require_parse = True
 
     def read_all(self, paths: Sequence[PurePath]) -> None:
@@ -279,18 +285,22 @@ class Stylesheet:
         for path in paths:
             self.read(path)
 
-    def has_source(self, path: str | PurePath) -> bool:
+    def has_source(self, path: str, class_var: str = "") -> bool:
         """Check if the stylesheet has this CSS source already.
+
+        Args:
+            path: The file path of the source in question.
+            class_var: The widget class variable we might be reading the CSS from.
 
         Returns:
             Whether the stylesheet is aware of this CSS source or not.
         """
-        return str(path) in self.source
+        return (path, class_var) in self.source
 
     def add_source(
         self,
         css: str,
-        path: str | PurePath | None = None,
+        read_from: CSSLocation | None = None,
         is_default_css: bool = False,
         tie_breaker: int = 0,
         scope: str = "",
@@ -299,6 +309,7 @@ class Stylesheet:
 
         Args:
             css: String with CSS source.
+            read_from: The original source location of the CSS.
             path: The path of the source if a file, or some other identifier.
             is_default_css: True if the CSS is defined in the Widget, False if the CSS is defined
                 in a user stylesheet.
@@ -310,17 +321,18 @@ class Stylesheet:
             StylesheetParseError: If the CSS is invalid.
         """
 
-        if path is None:
-            path = str(hash(css))
-        elif isinstance(path, PurePath):
-            path = str(css)
-        if path in self.source and self.source[path].content == css:
-            # Path already in source, and CSS is identical
-            content, is_defaults, source_tie_breaker, scope = self.source[path]
+        if read_from is None:
+            read_from = ("", str(hash(css)))
+
+        if read_from in self.source and self.source[read_from].content == css:
+            # Location already in source and CSS is identical.
+            content, is_defaults, source_tie_breaker, scope = self.source[read_from]
             if source_tie_breaker > tie_breaker:
-                self.source[path] = CssSource(content, is_defaults, tie_breaker, scope)
+                self.source[read_from] = CssSource(
+                    content, is_defaults, tie_breaker, scope
+                )
             return
-        self.source[path] = CssSource(css, is_default_css, tie_breaker, scope)
+        self.source[read_from] = CssSource(css, is_default_css, tie_breaker, scope)
         self._require_parse = True
 
     def parse(self) -> None:
@@ -332,13 +344,18 @@ class Stylesheet:
         rules: list[RuleSet] = []
         add_rules = rules.extend
 
-        for path, (css, is_default_rules, tie_breaker, scope) in self.source.items():
+        for read_from, (
+            css,
+            is_default_rules,
+            tie_breaker,
+            scope,
+        ) in self.source.items():
             if css in self._invalid_css:
                 continue
             try:
                 css_rules = self._parse_rules(
                     css,
-                    path,
+                    read_from=read_from,
                     is_default_rules=is_default_rules,
                     tie_breaker=tie_breaker,
                     scope=scope,
@@ -364,10 +381,10 @@ class Stylesheet:
         """
         # Do this in a fresh Stylesheet so if there are errors we don't break self.
         stylesheet = Stylesheet(variables=self._variables)
-        for path, (css, is_defaults, tie_breaker, scope) in self.source.items():
+        for read_from, (css, is_defaults, tie_breaker, scope) in self.source.items():
             stylesheet.add_source(
                 css,
-                path,
+                read_from=read_from,
                 is_default_css=is_defaults,
                 tie_breaker=tie_breaker,
                 scope=scope,